Allows workflow to run on all branches
Previously, the build and release workflow was restricted to triggering only on `main` branch pushes and pull requests.

This change removes the branch filter, enabling the workflow to execute on pushes and pull requests to any branch. This provides earlier validation and feedback for feature branches and ensures a more comprehensive CI/CD process.

Signed-off-by: Banee Ishaque K <baneeishaque@gmail.com>