Use vfox to install Flutter on Windows
Replaces the workaround script for installing Flutter on Windows with a direct vfox installation.
This change removes the need for the `install-flutter-windows-workaround.ps1` script, which was implemented due to issues with the vfox backend on Windows.
The new script `install-flutter-vfox.ps1` leverages vfox directly to add and use the specified Flutter version.
The workflow `build-and-release.yml` is updated to use the new script and also includes the installation of the vfox CLI via winget.
The `MISE_DISABLE_BACKENDS` environment variable is set to `vfox` in the `build-windows` job to ensure mise doesn't interfere with the vfox installation.

Signed-off-by: Banee Ishaque K <baneeishaque@gmail.com>